+++ test/libsolidity/syntaxTests/using/external_function_qualified_with_this.sol
@@ -1,0 +1,7 @@
+contract C {
+    using {this.contractFunction} for uint;
+
+    function contractFunction(uint) external view {}
+}
+// ----
+// DeclarationError 9589: (24-45): Function name expected, but the identifier was not found or is not unique.