+++ test/libsolidity/syntaxTests/using/free_functions_non_unique_err.sol
@@ -9,4 +9,4 @@ contract C {
     using {id} for uint256;
 }
 // ----
-// DeclarationError 7920: (145-147): Identifier not found or not unique.
+// DeclarationError 9589: (145-147): Function name expected, but the identifier was not found or is not unique.