+++ test/libsolidity/syntaxTests/using/free_overloads.sol
@@ -7,4 +7,4 @@ function f(int8 storage x) pure returns (int) {
 using {f} for uint8;
 using {f} for int;
 // ----
-// DeclarationError 7920: (132-133): Identifier not found or not unique.
+// DeclarationError 9589: (132-133): Function name expected, but the identifier was not found or is not unique.