+++ test/libsolidity/syntaxTests/using/free_overloads_array.sol
@@ -6,4 +6,4 @@ function f(uint x, uint y) pure returns (int) {
 }
 using {f} for uint;
 // ----
-// DeclarationError 7920: (138-139): Identifier not found or not unique.
+// DeclarationError 9589: (138-139): Function name expected, but the identifier was not found or is not unique.