+++ test/libsolidity/syntaxTests/using/function_from_base_contract_qualified_with_super.sol
@@ -1,0 +1,9 @@
+contract C {
+    function baseFunction(uint) public pure {}
+}
+
+contract D is C {
+    using {super.baseFunction} for uint;
+}
+// ----
+// DeclarationError 9589: (92-110): Function name expected, but the identifier was not found or is not unique.