+++ test/libsolidity/syntaxTests/using/undeclared_library.sol
@@ -1,0 +1,3 @@
+using L for uint;
+// ----
+// DeclarationError 9589: (6-7): Library name expected, but the identifier was not found or is not unique.